2nd WA abc214c.scala

+++ ABC/201-250/abc214/C.scala
@@ -2,10 +2,10 @@ import scala.io.StdIn._
 
 object Main extends App {
   val n = readInt()
-  val s = readLine().split(" ").map(_.toInt)
-  val t = readLine().split(" ").map(_.toInt)
+  val s = readLine().split(" ").map(_.toLong)
+  val t = readLine().split(" ").map(_.toLong)
 
-  val ans = new Array[Int](n)
+  val ans = new Array[Long](n)
 
   ans(0) = math.min(s(0), t(0))
   (1 until n).foreach { i =>
